web/requirements.txt
Django>=4.0,<5.0
django-tinymce==3.4.0
django-simple-history==3.0.0
django-model-utils==4.1.1